Dock: make sim time display more precise. Fix portability issue with FlightProfile.
FlightProfile was modified to run within the brata.station framework, but that broke its ability
to run as a standalone program.  That capability has been restored.  However, FlightProfile has
a dependency on station.state.py now, which requires that the dir containing station/state.py
be in PYTHONPATH.  This should be fixed.